Major refactoring of service

+++ livefeedback_hub/handlers/manage.py
@@ -1,45 +1,26 @@
-import logging
-import os
 import shutil
+import subprocess
 import tempfile
+import unittest.mock
 import uuid
+import zipfile
 from concurrent.futures.thread import ThreadPoolExecutor
-from hashlib import md5
 from io import BytesIO
 from typing import Optional
-import pkg_resources
 
-from python_on_whales import docker
-from python_on_whales.exceptions import NoSuchImage
-import stdio_proxy
+import pkg_resources
 from jupyterhub.services.auth import HubOAuthenticated
-from otter.grade import utils, containers
+from otter.grade import utils
+from python_on_whales import docker
 from tornado import web
 from tornado.httputil import HTTPFile
-from tornado.web import RequestHandler
 
 from livefeedback_hub import core
 from livefeedback_hub.core import teacher_only
-from livefeedback_hub.db import AutograderZip, Result
+from livefeedback_hub.db import AutograderZip, Result, State
 from livefeedback_hub.server import JupyterService
 
-executor = ThreadPoolExecutor(max_workers=16)
-
-
-def delete_docker_image(service: JupyterService, task: AutograderZip):
-    """
-    Trys to delete the docker image belonging to the provided task
-    :param service: a service instance used for logging
-    :param task: the task to delete
-    """
-    m = md5()
-    m.update(task.data)
-    image = f"{utils.OTTER_DOCKER_IMAGE_TAG}:{m.hexdigest()}"
-    service.log.info(f"Deleting docker image {image}")
-    try:
-        docker.image.remove(image, force=True)
-    except NoSuchImage as e:
-        service.log.warning(f"Image not found: {e}")
+manage_executor = ThreadPoolExecutor(max_workers=16)
 
 
 def build(service: JupyterService, id: str, zip_file: HTTPFile, update: bool = False):
@@ -50,9 +31,6 @@ def build(service: JupyterService, id: str, zip_file: HTTPFile, update: bool = F
     :param zip_file: the provided zip file
     :param update: flag indicating whether an update is executed (or a new image was added)
     """
-    tmp_dir = tempfile.mkdtemp()
-    fd_zip, path_zip = tempfile.mkstemp(suffix=".zip", dir=tmp_dir)
-    cwd = os.getcwd()
 
     with service.session() as session:
         item: Optional[AutograderZip] = session.query(AutograderZip).filter_by(id=id).first()
@@ -60,48 +38,42 @@ def build(service: JupyterService, id: str, zip_file: HTTPFile, update: bool = F
             return
         try:
 
-            with os.fdopen(fd_zip, "wb") as tmp:
-                tmp.write(zip_file["body"])
-                tmp.flush()
-
-            os.chdir(tmp_dir)
             base = "ucbdsinfra/otter-grader"
             service.log.info(f"Building new docker image for {id}")
-            image = utils.OTTER_DOCKER_IMAGE_TAG + ":" + containers.generate_hash(os.path.basename(path_zip))
+            image = utils.OTTER_DOCKER_IMAGE_TAG + ":" + core.calcuate_zip_hash(zip_file["body"])
 
             if update and docker.image.exists(image):
                 service.log.info(f"Image for {id} exists ({image})")
-                item.ready = True
+                item.state = State.ready
                 return
             dockerfile = pkg_resources.resource_filename("otter.grade", "Dockerfile")
 
             if not docker.image.exists(image):
-                service.log.info(f"Building new image for {id} using {base} as base image")
-                stdout = BytesIO()
-                stderr = BytesIO()
-                with stdio_proxy.redirect_stdout(stdout), stdio_proxy.redirect_stderr(stderr):
-                    docker.build(".", build_args={"ZIPPATH": path_zip, "BASE_IMAGE": base}, tags=[image], file=dockerfile, load=True)
-                service.log.info(f"Building new docker image {image} for {id} completed")
+                with tempfile.TemporaryDirectory() as tmp_dir:
+                    with zipfile.ZipFile(BytesIO(zip_file["body"]), "r") as zip_ref:
+                        zip_ref.extractall(tmp_dir)
+                    shutil.copy(dockerfile, tmp_dir)
+                    service.log.info(f"Building new image for {id} using {base} as base image")
+                    run = core.timeout_injector(subprocess.run)
+                    with unittest.mock.patch("subprocess.run", run):
+                        for line in docker.build(tmp_dir, build_args={"BASE_IMAGE": base}, tags=[image], file=dockerfile, load=True, stream_logs=True):
+                            service.log.debug(line)
+                    service.log.info(f"Building new docker image {image} for {id} completed")
 
         except Exception as e:
             service.log.error(f"Error while building docker image for {id}: {e}")
-            raise
+            item.state = State.error
+            return
 
-        finally:
-            os.chdir(cwd)
-            shutil.rmtree(tmp_dir)
-
-        if update:
-            delete_docker_image(service, item)
+        if update and core.calcuate_zip_hash(zip_file["body"]) != core.calcuate_zip_hash(item.data):
+            core.delete_docker_image(service, item)
+        service.log.info(f"Marking {id} as ready")
         item.data = zip_file["body"]
-        item.ready = True
+        item.state = State.ready
+        session.commit()
 
 
-class FeedbackManagementHandler(HubOAuthenticated, RequestHandler):
-    def initialize(self, service: JupyterService):
-        self.service = service
-        self.log: logging.Logger = service.log
-
+class FeedbackManagementHandler(HubOAuthenticated, core.CustomRequestHandler):
     @teacher_only
     async def get(self):
         user_hash = core.get_user_hash(self.get_current_user())
@@ -111,11 +83,7 @@ class FeedbackManagementHandler(HubOAuthenticated, RequestHandler):
             await self.render("overview.html", tasks=tasks, base=self.service.prefix)
 
 
-class FeedbackZipAddHandler(HubOAuthenticated, RequestHandler):
-    def initialize(self, service: JupyterService):
-        self.service = service
-        self.log: logging.Logger = service.log
-
+class FeedbackZipAddHandler(HubOAuthenticated, core.CustomRequestHandler):
     @teacher_only
     async def get(self):
         task = AutograderZip()
@@ -138,17 +106,13 @@ class FeedbackZipAddHandler(HubOAuthenticated, RequestHandler):
         with self.service.session() as session:
             # get new uuid
             new_uuid = str(uuid.uuid4())
-            item = AutograderZip(id=new_uuid, owner=user_hash, data=zip_file["body"], description=description, ready=False)
+            item = AutograderZip(id=new_uuid, owner=user_hash, data=zip_file["body"], description=description, state=State.building)
             session.add(item)
             session.commit()
-            executor.submit(build, self.service, new_uuid, zip_file)
+            manage_executor.submit(build, self.service, new_uuid, zip_file)
 
 
-class FeedbackZipDeleteHandler(HubOAuthenticated, RequestHandler):
-    def initialize(self, service: JupyterService):
-        self.service = service
-        self.log: logging.Logger = service.log
-
+class FeedbackZipDeleteHandler(HubOAuthenticated, core.CustomRequestHandler):
     @teacher_only
     async def get(self, live_id: str):
 
@@ -158,19 +122,19 @@ class FeedbackZipDeleteHandler(HubOAuthenticated, RequestHandler):
             if not task:
                 raise web.HTTPError(403)
             else:
-                # Delete task from database and delete docker image
+
+                if task.state == State.building:
+                    await self.render("error.html", base=self.service.prefix)
+                    return
+                    # Delete task from database and delete docker image
                 self.service.log.info(f"Deleting task {live_id}")
-                delete_docker_image(self.service, task)
+                core.delete_docker_image(self.service, task)
                 session.delete(task)
             session.query(Result).filter_by(assignment=live_id).delete()
         self.redirect(self.service.prefix)
 
 
-class FeedbackZipUpdateHandler(HubOAuthenticated, RequestHandler):
-    def initialize(self, service: JupyterService):
-        self.service = service
-        self.log: logging.Logger = service.log
-
+class FeedbackZipUpdateHandler(HubOAuthenticated, core.CustomRequestHandler):
     @teacher_only
     async def get(self, live_id: str):
 
@@ -180,7 +144,10 @@ class FeedbackZipUpdateHandler(HubOAuthenticated, RequestHandler):
             if not task:
                 raise web.HTTPError(403)
             else:
-                await self.render("edit.html", task=task, edit=True, base=self.service.prefix)
+                if task.state == State.building:
+                    await self.render("error.html", base=self.service.prefix)
+                else:
+                    await self.render("edit.html", task=task, edit=True, base=self.service.prefix)
 
     @teacher_only
     async def post(self, live_id: str):
@@ -191,6 +158,9 @@ class FeedbackZipUpdateHandler(HubOAuthenticated, RequestHandler):
             if not task:
                 raise web.HTTPError(403)
             else:
+                if task.state == State.building:
+                    await self.render("error.html", base=self.service.prefix)
+                    return
                 task.description = self.get_body_argument("description")
                 session.commit()
 
@@ -207,6 +177,6 @@ class FeedbackZipUpdateHandler(HubOAuthenticated, RequestHandler):
         with self.service.session() as session:
             task: Optional[AutograderZip] = session.query(AutograderZip).filter_by(id=live_id, owner=user_hash).first()
             # Mark as not ready, rebuild image and mark as ready afterwards
-            task.ready = False
+            task.state = State.building
             session.commit()
-            executor.submit(build, self.service, live_id, zip_file, update=True)
+            manage_executor.submit(build, self.service, live_id, zip_file, update=True)
